+++ jest.config.js
@@ -42,8 +42,8 @@ module.exports = {
     // '<rootDir>/src/components/toaster/index.test.js',
     // '<rootDir>/src/components/votesPreview/votesPreview.test.js',
     // '<rootDir>/src/components/accountVisual/index.test.js',
-    // '<rootDir>/src/components/sendReadable/index.test.js',
-    // '<rootDir>/src/components/sendReadable/send.test.js',
+    // '<rootDir>/src/components/send/steps/confirm/index.test.js',
+    // '<rootDir>/src/components/send/steps/confim/confirm.test.js',
     // '<rootDir>/src/components/box/index.test.js',
     // '<rootDir>/src/components/app/index.test.js',
     // '<rootDir>/src/components/secondPassphrase/index.test.js',
@@ -115,8 +115,8 @@ module.exports = {
     // '<rootDir>/src/components/toolbox/transitionWrapper/index.test.js',
     // '<rootDir>/src/components/request/specifyRequest.test.js',
     // '<rootDir>/src/components/request/confirmRequest.test.js',
-    // '<rootDir>/src/components/sendWritable/index.test.js',
-    // '<rootDir>/src/components/sendWritable/send.test.js',
+    // '<rootDir>/src/components/send/steps/form/index.test.js',
+    // '<rootDir>/src/components/send/steps/form/form.test.js',
     // '<rootDir>/src/components/actionBar/index.test.js',
     // '<rootDir>/src/components/mainMenu/mainMenu.test.js',
     // '<rootDir>/src/components/mainMenu/index.test.js',
