+++ src/components/send/steps/confirm/confirm.css
@@ -1,4 +1,4 @@
-@import '../app/variables.css';
+@import '../../../app/variables.css';
 
 :root {
   --tab-inactive-color: --color-grayscale-dark;
