+++ src/components/send/steps/confirm/index.test.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
-import i18n from '../../i18n';
-import SendHOC from './index';
+import i18n from '../../../../i18n';
+import ConfirmHOC from './index';
 
-describe('SendReadableHOC', () => {
+describe('ConfirmHOC', () => {
   let wrapper;
   const store = {};
   const account = {};
@@ -17,15 +17,15 @@ describe('SendReadableHOC', () => {
     });
     store.subscribe = () => {};
     store.dispatch = () => {};
-    wrapper = mount(<Provider store={store}><SendHOC i18n={i18n} /></Provider>);
+    wrapper = mount(<Provider store={store}><ConfirmHOC i18n={i18n} /></Provider>);
   });
 
   it('should render Send', () => {
-    expect(wrapper.find('SendReadable')).to.have.lengthOf(1);
+    expect(wrapper.find('Confirm')).to.have.lengthOf(1);
   });
 
   it('should mount Send with appropriate properties', () => {
-    const props = wrapper.find('SendReadable').props();
+    const props = wrapper.find('Confirm').props();
     expect(props.account).to.be.equal(account);
     expect(typeof props.sent).to.be.equal('function');
   });
