+++ src/components/send/steps/confirm/input.css
