+++ src/components/send/steps/form/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { translate } from 'react-i18next';
 
-import Send from './send';
+import Form from './form';
 
 const mapStateToProps = state => ({
   account: state.account,
@@ -9,4 +9,4 @@ const mapStateToProps = state => ({
   followedAccounts: state.followedAccounts.accounts,
 });
 
-export default connect(mapStateToProps)(translate()(Send));
+export default connect(mapStateToProps)(translate()(Form));
